reafactor(CompaniesValidation): rename files
