+++ frontend/src/pages/backoffice/CompaniesValidation/CompaniesRejectedTable.jsx
