+++ frontend/src/pages/backoffice/CompaniesValidation/CompaniesValidation.module.scss
@@ -1,4 +1,4 @@
-.CompanyValidation {
+.CompaniesValidation {
   &Modal {
     height: 100%;
     display: grid;
